django/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-role-permissions==3.2.0

pillow==10.4.0
validate-docbr==1.10.0

requests==2.32.3
psycopg2-binary>=2.9.6,<2.10



